+++ lib/l10n/app_localizations_en.dart
@@ -925,6 +925,17 @@ class AppLocalizationsEn extends AppLocalizations {
   }
 
   @override
+  String mapEntriesHere(int count) {
+    String _temp0 = intl.Intl.pluralLogic(
+      count,
+      locale: localeName,
+      other: '$count entries here',
+      one: '1 entry here',
+    );
+    return '$_temp0';
+  }
+
+  @override
   String get mapColor => 'Color on the map';
 
   @override